requirements.txt
streamlit
crewai
crewai_tools
python-dotenv
langchain-community
python-docx